perf(webui): keep the API client out of the shared app-shell bundle
Every route shipped ~690 KB gzip of JavaScript with almost no variation
between routes: the shared _app chunk was 680 KB and route-specific code
was 1-15 KB, so /login downloaded essentially the whole application.

One import chain was responsible for a large part of it:

  _app.tsx
    -> ServerReadinessGate.tsx
      -> store/connection.tsx
        -> services/tldw-server.ts        (subscribes at module scope)
          -> services/tldw/index.ts        (barrel, re-exports everything)
            -> services/tldw/TldwApiClient.ts + all domain modules

connection.tsx needed exactly one function from tldw-server.ts,
getStoredTldwServerURL, but tldw-server.ts registers a model-cache
invalidation subscription at module scope, so it cannot be imported
lazily. Extract that function into services/tldw-server-url.ts, which
depends only on the settings storage helper, and point connection.tsx at
it. tldw-server.ts re-exports the function so other importers are
unaffected.

Also make connection.tsx load tldwClient on first use rather than
statically; all 14 call sites were already in async context.

Measured on a production build:

  shared _app: 679.7 -> 558.0 KB gzip  (2.30 -> 1.86 MB raw)
  /login:      694.9 -> 573.2 KB gzip
  domain modules (admin, media, presentations, characters) no longer
  appear in the shared chunk at all

Two store tests mocked @/services/tldw-server for this function and now
mock the new module.
