+++ apps/packages/ui/src/store/connection.tsx
@@ -1,7 +1,13 @@
 import { createWithEqualityFn } from "zustand/traditional"
 
-import { tldwClient, type TldwConfig } from "@/services/tldw/TldwApiClient"
-import { getStoredTldwServerURL } from "@/services/tldw-server"
+import type { TldwConfig } from "@/services/tldw/TldwApiClient"
+
+// Loaded on first use rather than statically imported. This module is reachable
+// from _app, and a static import here pulled TldwApiClient -- and every API domain
+// module it re-exports -- into the shared bundle that every page downloads.
+const getTldwClient = async () =>
+  (await import("@/services/tldw/TldwApiClient")).tldwClient
+import { getStoredTldwServerURL } from "@/services/tldw-server-url"
 import { apiSend } from "@/services/api-send"
 import { createSafeStorage } from "@/utils/safe-storage"
 import {
@@ -240,7 +246,7 @@ const setUserPersonaFlag = async (persona: UserPersona): Promise<void> => {
 
 const ensurePlaceholderConfig = async (): Promise<string | null> => {
   try {
-    const cfg = await tldwClient.getConfig()
+    const cfg = await (await getTldwClient()).getConfig()
     if (cfg?.serverUrl) return cfg.serverUrl
   } catch {
     // ignore missing config
@@ -248,7 +254,7 @@ const ensurePlaceholderConfig = async (): Promise<string | null> => {
 
   const placeholderUrl = "http://127.0.0.1:0"
   try {
-    await tldwClient.updateConfig({
+    await (await getTldwClient()).updateConfig({
       serverUrl: placeholderUrl,
       authMode: "single-user",
       apiKey: "test-bypass"
@@ -527,7 +533,7 @@ const initialState: ConnectionState = {
 
 const getPersistedServerUrl = async (): Promise<string | null> => {
   try {
-    const cfg = await tldwClient.getConfig()
+    const cfg = await (await getTldwClient()).getConfig()
     const quickstartWebUiServerUrl = getQuickstartWebUiServerUrl()
     if (quickstartWebUiServerUrl) {
       return quickstartWebUiServerUrl
@@ -739,7 +745,7 @@ export const useConnectionStore = createWithEqualityFn<ConnectionStore>((set, ge
       }))
 
       try {
-        let cfg = await tldwClient.getConfig()
+        let cfg = await (await getTldwClient()).getConfig()
         const quickstartWebUiServerUrl = getQuickstartWebUiServerUrl()
         const recoveryProbeSourceServerUrl = cfg?.serverUrl ?? currentState.serverUrl ?? null
         let serverUrl = quickstartWebUiServerUrl ?? cfg?.serverUrl ?? null
@@ -748,7 +754,7 @@ export const useConnectionStore = createWithEqualityFn<ConnectionStore>((set, ge
           quickstartWebUiServerUrl &&
           cfg?.serverUrl !== quickstartWebUiServerUrl
         ) {
-          await tldwClient.updateConfig({ serverUrl: quickstartWebUiServerUrl })
+          await (await getTldwClient()).updateConfig({ serverUrl: quickstartWebUiServerUrl })
           cfg = {
             ...(cfg || {}),
             serverUrl: quickstartWebUiServerUrl,
@@ -763,10 +769,10 @@ export const useConnectionStore = createWithEqualityFn<ConnectionStore>((set, ge
             // fall back to the hard-coded localhost default here.
             const storedUrl = await getStoredTldwServerURL()
             if (storedUrl) {
-              await tldwClient.updateConfig({
+              await (await getTldwClient()).updateConfig({
                 serverUrl: storedUrl
               })
-              cfg = await tldwClient.getConfig()
+              cfg = await (await getTldwClient()).getConfig()
               serverUrl = cfg?.serverUrl ?? storedUrl
             }
           } catch {
@@ -832,7 +838,7 @@ export const useConnectionStore = createWithEqualityFn<ConnectionStore>((set, ge
           return
         }
 
-        await tldwClient.initialize()
+        await (await getTldwClient()).initialize()
 
         // Request health via background for detailed status codes.
         // Health endpoints may require auth; apiSend injects headers based
@@ -881,7 +887,7 @@ export const useConnectionStore = createWithEqualityFn<ConnectionStore>((set, ge
           )
           if (probeOk) {
             if (!quickstartWebUiServerUrl) {
-              await tldwClient.updateConfig({ serverUrl: fallbackServerUrl })
+              await (await getTldwClient()).updateConfig({ serverUrl: fallbackServerUrl })
               serverUrl = fallbackServerUrl
               cfg = {
                 ...(cfg || {}),
@@ -926,7 +932,7 @@ export const useConnectionStore = createWithEqualityFn<ConnectionStore>((set, ge
           try {
             // Add timeout to RAG health check to prevent hanging
             // Increased from 5s to 15s to avoid false "offline" status when RAG is slow but working
-            const ragPromise = tldwClient.ragHealth()
+            const ragPromise = (await getTldwClient()).ragHealth()
             const ragTimeout = new Promise<null>((resolve) =>
               setTimeout(() => resolve(null), 15000)
             )
@@ -1059,7 +1065,7 @@ export const useConnectionStore = createWithEqualityFn<ConnectionStore>((set, ge
   },
 
   async setServerUrl(url: string) {
-    await tldwClient.updateConfig({ serverUrl: url })
+    await (await getTldwClient()).updateConfig({ serverUrl: url })
     await get().checkOnce()
   },
 
@@ -1079,7 +1085,7 @@ export const useConnectionStore = createWithEqualityFn<ConnectionStore>((set, ge
       prev.mode === "demo" || prev.offlineBypass === true
     let config: TldwConfig | null = null
     try {
-      config = await tldwClient.getConfig()
+      config = await (await getTldwClient()).getConfig()
     } catch {
       // ignore config lookup failures and fall back to current state
     }
@@ -1129,7 +1135,7 @@ export const useConnectionStore = createWithEqualityFn<ConnectionStore>((set, ge
 
   async restartOnboarding() {
     const prev = get().state
-    await tldwClient.clearManualSingleUserCredentials()
+    await (await getTldwClient()).clearManualSingleUserCredentials()
     await setFirstRunCompleteFlag(false)
     set({
       state: {
@@ -1153,7 +1159,7 @@ export const useConnectionStore = createWithEqualityFn<ConnectionStore>((set, ge
   },
 
   async setConfigPartial(config: Partial<TldwConfig>) {
-    await tldwClient.updateConfig(config)
+    await (await getTldwClient()).updateConfig(config)
     const prev = get().state
 
     let nextStep: ConnectionState["configStep"] = prev.configStep
